Resolved crash when clicking on a Sil-Q X11 window
XChangeProperty:format specifies whether the data should be viewed as 8,16,32 bit.
The data in this case is an Atom, which is a typedef for unsigned long on x64 (64bits).
This causes a crash, since 64 bit is not allowed. While this is confusing, 32 bit
should be used even though the data itself might be 64 bit.