Changes player tile depending on race.
Note: This is a little fragile, as it forces races to have a representation
in monster.txt and the tileset.

When adding a new race, you should update monster.txt and graf-new.prf.
The race number used in both (R) should match the race number (N) used
in race.txt.