Alpha 0.5 release!
Finally a closed alpha release with main ideas implemented! Omg, that took me so long. Now lot's of testing, figuring bugs (multiples I guess) and so on and so forth until 0.6. But I'm proud I did at least this, and it should be at a playable stage. We will see how it goes further.